Rename EricaAuftragRepository -> EricaRequestRepository
