+++ erica/application/EricaAuftrag/EricaAuftragService.py
@@ -5,7 +5,7 @@ from opyoid import Injector, Module
 
 from erica.infrastructure.InfrastructureModule import InfrastructureModule
 from erica.infrastructure.rq.RqModule import RqModule
-from erica.infrastructure.sqlalchemy.repositories.EricaAuftragRepository import EricaAuftragRepository
+from erica.infrastructure.sqlalchemy.repositories.erica_request_repository import EricaRequestRepository
 
 injector = Injector([InfrastructureModule(), RqModule()])
 
@@ -19,9 +19,9 @@ class EricaAuftragServiceInterface:
 
 
 class EricaAuftragService(EricaAuftragServiceInterface):
-    erica_auftrag_repository: EricaAuftragRepository
+    erica_auftrag_repository: EricaRequestRepository
 
-    def __init__(self, repository: EricaAuftragRepository = injector.inject(EricaAuftragRepository)) -> None:
+    def __init__(self, repository: EricaRequestRepository = injector.inject(EricaRequestRepository)) -> None:
         super().__init__()
         self.erica_auftrag_repository = repository
