+++ erica/domain/Repositories/EricaRequestRepositoryInterface.py
@@ -4,5 +4,5 @@ from erica.domain.EricaAuftrag.EricaAuftrag import EricaAuftrag
 from erica.domain.Repositories.BaseRepositoryInterface import BaseRepositoryInterface
 
 
-class EricaAuftragRepositoryInterface(BaseRepositoryInterface[EricaAuftrag], ABC):
+class EricaRequestRepositoryInterface(BaseRepositoryInterface[EricaAuftrag], ABC):
     pass
