+++ erica/infrastructure/sqlalchemy/repositories/erica_request_repository.py
@@ -4,14 +4,14 @@ import injector as injector
 from sqlalchemy.orm import Session
 
 from erica.domain.EricaAuftrag.EricaAuftrag import EricaAuftrag
-from erica.domain.Repositories.EricaAuftragRepositoryInterface import EricaAuftragRepositoryInterface
+from erica.domain.Repositories.EricaRequestRepositoryInterface import EricaRequestRepositoryInterface
 from erica.infrastructure.sqlalchemy.EricaAuftragSchema import EricaAuftragSchema
-from erica.infrastructure.sqlalchemy.repositories.BaseRepository import BaseRepository
+from erica.infrastructure.sqlalchemy.repositories.base_repository import BaseRepository
 
 
-class EricaAuftragRepository(
+class EricaRequestRepository(
     BaseRepository[EricaAuftrag, EricaAuftragSchema],
-    EricaAuftragRepositoryInterface,
+    EricaRequestRepositoryInterface,
     ABC
 ):
     def __init__(self, db_connection: Session = injector.inject(Session)):
